+++ src/device/mod.rs
@@ -5,8 +5,8 @@ use crate::device::mmio::MmioDeviceSpace;
 use crate::device::virtio::block::{self, BlockArgs};
 use crate::device::virtio::{CommonArgs, MmioConfig};
 use crate::kvm::hypervisor::{Hypervisor, VmMem};
-use crate::proc::Mapping;
 use crate::result::Result;
+use crate::tracer::proc::Mapping;
 use event_manager::{EventManager, MutEventSubscriber};
 use simple_error::{bail, try_with};
 use std::path::PathBuf;
