+++ src/kvm/fd_transfer.rs
@@ -7,10 +7,10 @@ use std::mem::size_of;
 use std::os::unix::prelude::*;
 use std::sync::{Arc, RwLock};
 
-use crate::inject_syscall;
 use crate::kvm::hypervisor::HvMem;
 use crate::kvm::tracee::Tracee;
 use crate::result::Result;
+use crate::tracer::inject_syscall;
 
 // inspired by https://github.com/Mic92/cntr/blob/492b2d9e9abc9ccd4f01a0134aab73df16393423/src/ipc.rs
 pub struct Socket {
